Create results.py
First upload of "results.py"! Currently this is a standalone file (I haven't looked at importing the other files to make the state and candidate tables yet.) I currently commented that out just to be clear about what the new code here is. Uncomment the top chunk and the main if you want to use it by itself just to see it work.